:recycle: [Refactor]: 오타 및 대소문자 수정
