+++ src/features/input-schedule/lib/index.ts
@@ -1,1 +1,1 @@
-export * from "./ScheduleFormTypes";
+export * from "./scheduleFormTypes";
