+++ src/features/input-schedule/lib/scheduleFormTypes.ts
